requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
spacy>=3.7.0
sentence-transformers>=2.2.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.17.0
reportlab>=4.0.0
numpy>=1.24.0
python-docx>=0.8.11
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
PyPDF2>=3.0.0
python-docx>=0.8.11
